+++ pkg/webui/console/views/application-integrations-webhook-edit/index.js
@@ -95,6 +95,7 @@ const webhookEntitySelector = [
   'template_ids',
   'health_status',
   'field_mask',
+  'paused',
 ]
 
 const ApplicationWebhookEdit = () => {