+++ CHANGELOG.md
@@ -15,6 +15,7 @@ For details about compatibility between different releases, see the **Commitment
   - This requires a database migration.
 - Support for managing MAC settings profiles.
   - This feature is experimental and subject to change.
+- Support pausing webhook in the edit webhook view in console
 
 ### Changed